# dc-api/openapi.yaml
openapi: 3.0.3
info:
  title: DC-API
  version: 0.1.0
  description: |
    WSO2 Sovereign Cloud Control Plane API.

    DC-API sits in front of a Rancher + Harvester infrastructure layer and
    exposes a clean, tenant-isolated REST API for managing VMs, RKE2 clusters,
    virtual networks, DNS zones, and RBAC — in the style of AWS / Azure.

    **Authentication**

    Every `/v1/*` endpoint requires a `Bearer` token in the `Authorization`
    header. Two token types are accepted by the same scheme:

    1. **Asgardeo OIDC JWT** — obtained via the `dcctl login` PKCE flow. The JWT
       must carry a `groups` claim containing at least one group matching the
       tenant group prefix (default `dc-tenant-<tenant-id>`).

    2. **Service-account token** — a long-lived opaque token with the format
       `dcapi_sa_<lookup_id>_<secret>`, issued by
       `POST /v1/tenants/{tenant_id}/service-accounts`. Store it on creation;
       it cannot be retrieved again.

    **Async provisioning**

    Long-running operations (VM create, VNet create, etc.) return `202 Accepted`
    immediately. The response body contains the resource object with
    `"status": "PENDING"` and a `note` field with the polling URL. Poll the
    GET endpoint until `status` transitions to `ACTIVE` or `FAILED`.

    **Status lifecycle**

    `PENDING → ACTIVE` or `PENDING → FAILED`

    Deletion: `ACTIVE → DELETING → (row removed on success)`

    **Error format**

...
servers:
...
  - url: http://localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /v1/tenants/{tenant_id}/projects/{project_id}/vnets/{vnet_id}/subnets/{subnet_id}:
    parameters:
      - $ref: "#/components/parameters/TenantID"
      - $ref: "#/components/parameters/ProjectID"
      - $ref: "#/components/parameters/VNetID"
      - name: subnet_id
        in: path
        required: true
        description: UUID of the subnet
        schema:
          type: string
          format: uuid
          example: 3a1c2e5d-0000-0000-0000-000000000021

    get:
      operationId: getSubnet
      tags: [networking-vnets]
      summary: Get a subnet by UUID
      responses:
        "200":
          description: Subnet resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subnet"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteSubnet
      tags: [networking-vnets]
      summary: Delete a subnet (async)
      description: |
        Marks the subnet DELETING and triggers async KubeOVN removal.

        Returns 409 if any VMs, bastions, or clusters are still attached to
        the subnet (via `subnet_id` on the resource row). KubeOVN refuses to
        release the logical switch while LSPs are pinned, so attempting the
        delete asynchronously would leave the subnet stuck in FAILED. Delete
        the attached resources first, then retry.

        **Required role**: `owner`.
      responses:
        "202":
          description: Deletion accepted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Subnet has attached resources (VMs, bastions, or clusters)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "subnet has attached resources; delete them first: chalitha-bastion (bastion), web-01 (vm)"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Route Tables ───────────────────────────────────────────────────────────────
  /v1/tenants/{tenant_id}/projects/{project_id}/vnets/{vnet_id}/route-tables:
    parameters:
      - $ref: "#/components/parameters/TenantID"
      - $ref: "#/components/parameters/ProjectID"
      - $ref: "#/components/parameters/VNetID"

    post:
      operationId: createRouteTable
      tags: [networking-route-tables]
      summary: Create a route table within a VNet (synchronous)
      description: |
        Synchronous (returns `201 Created`). Writes static routes onto the
        KubeOVN Vpc CRD immediately.

        In M2 all routes apply at the VPC level (per-subnet differentiation
        is informational only — see `/associations`).

        The VNet must be `ACTIVE`.

        **Required role**: `member` or `owner`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRouteTableRequest"
            example:
              name: rt-default
              description: Default route table
              routes:
                - name: internet-egress
                  destination_cidr: "0.0.0.0/0"
                  next_hop_type: internet
      responses:
        "201":
          description: Route table created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteTable"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      operationId: listRouteTables
      tags: [networking-route-tables]
      summary: List route tables within a VNet
      responses:
        "200":
          description: List of route tables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteTable"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/tenants/{tenant_id}/projects/{project_id}/vnets/{vnet_id}/route-tables/{rt_id}:
    parameters:
      - $ref: "#/components/parameters/TenantID"
      - $ref: "#/components/parameters/ProjectID"
      - $ref: "#/components/parameters/VNetID"
      - name: rt_id
        in: path
        required: true
        description: UUID of the route table
        schema:
          type: string
          format: uuid
          example: 3a1c2e5d-0000-0000-0000-000000000031

    get:
      operationId: getRouteTable
      tags: [networking-route-tables]
      summary: Get a route table by UUID
      responses:
        "200":
          description: Route table resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteTable"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: updateRouteTableRoutes
      tags: [networking-route-tables]
      summary: Replace all routes in a route table (synchronous)
      description: |
        Replaces the entire route set and writes it to the KubeOVN Vpc CRD
        immediately. Returns the updated route table with `200 OK`.

        **Required role**: `member` or `owner`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [routes]
              properties:
                routes:
                  type: array
                  items:
                    $ref: "#/components/schemas/RouteRule"
      responses:
        "200":
          description: Updated route table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteTable"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      operationId: deleteRouteTable
      tags: [networking-route-tables]
      summary: Delete a route table (synchronous)
      description: |
        Removes routes from the KubeOVN Vpc CRD and deletes the DB row.
        Returns `204 No Content`.

        **Required role**: `owner`.
      responses:
        "204":
          description: Route table deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/tenants/{tenant_id}/projects/{project_id}/vnets/{vnet_id}/route-tables/{rt_id}/associations:
    parameters:
      - $ref: "#/components/parameters/TenantID"
      - $ref: "#/components/parameters/ProjectID"
      - $ref: "#/components/parameters/VNetID"
      - name: rt_id
        in: path
        required: true
        description: UUID of the route table
        schema:
          type: string
          format: uuid
          example: 3a1c2e5d-0000-0000-0000-000000000031

    post:
      operationId: associateRouteTable
      tags: [networking-route-tables]
      summary: Associate a route table with a subnet (informational in M2)
      description: |
        Records the association in the DB and returns `201 Created` with a
        `warning` field indicating that per-subnet routing is informational
        only in M2 — all routes apply at the VPC level. Per-subnet OVN
        policy routes are slated for M2.5.

        **Required role**: `member` or `owner`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subnet_id]
              properties:
                subnet_id:
                  type: string
                  format: uuid
                  description: UUID of the subnet to associate
                  example: 3a1c2e5d-0000-0000-0000-000000000021
      responses:
        "201":
          description: Association created (informational in M2)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteTableAssociation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/tenants/{tenant_id}/projects/{project_id}/vnets/{vnet_id}/route-tables/{rt_id}/associations/{assoc_id}:
    parameters:
      - $ref: "#/components/parameters/TenantID"
      - $ref: "#/components/parameters/ProjectID"
      - $ref: "#/components/parameters/VNetID"
      - name: rt_id
        in: path
        required: true
        description: UUID of the route table
        schema:
          type: string
          format: uuid
      - name: assoc_id
        in: path
        required: true
        description: UUID of the association
        schema:
          type: string
          format: uuid
          example: 3a1c2e5d-0000-0000-0000-000000000041

    delete:
      operationId: disassociateRouteTable
      tags: [networking-route-tables]
      summary: Remove a route table–subnet association
      description: |
        Removes the association record. Returns `204 No Content`.

        **Required role**: `owner`.
      responses:
        "204":
          description: Association removed
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Peerings ───────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Accepts either:
        1. An Asgardeo OIDC JWT (RS256) — obtained via the PKCE flow (`dcctl login`).
        2. A DC-API service-account token with the format
           `dcapi_sa_<lookup_id>_<secret>` — issued by
           `POST /v1/tenants/{tenant_id}/service-accounts`.
  parameters:
...
    TenantID:
      name: tenant_id
      in: path
      required: true
      description: Tenant identifier (slug, e.g. "iam-platform")
      schema:
        type: string
        example: iam-platform
    ProjectID:
      name: project_id
      in: path
      required: true
      description: Project identifier (slug, e.g. "platform-infra")
      schema:
        type: string
        example: platform-infra
    VNetID:
      name: vnet_id
      in: path
      required: true
      description: UUID of the virtual network
      schema:
        type: string
        format: uuid
        example: 3a1c2e5d-0000-0000-0000-000000000010
  responses:
    BadRequest:
      description: Bad request — validation error or malformed JSON
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "name is required"
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "missing Authorization header"
    Forbidden:
      description: Insufficient role for this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "role 'member' is insufficient; 'owner' required"
...
    NotFound:
      description: Resource not found (also returned for cross-tenant access attempts)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "resource not found"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "failed to register resource"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable description of what went wrong
          example: "name is required"
...
    Subnet:
      type: object
      required: [id, vnet_id, tenant_id, name, cidr, status, provider_type, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          example: 3a1c2e5d-0000-0000-0000-000000000021
        vnet_id:
          type: string
          format: uuid
          example: 3a1c2e5d-0000-0000-0000-000000000010
        tenant_id:
          type: string
          example: iam-platform
        name:
          type: string
          example: app-subnet
        cidr:
          type: string
          example: 10.10.1.0/24
        gateway:
          type: string
          example: 10.10.1.1
        description:
          type: string
        status:
          $ref: "#/components/schemas/ResourceStatus"
        provider_type:
          type: string
          example: kubeovn
        message:
          type: string
        created_at:
          type: string
          format: date-time
          example: "2026-05-09T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-05-09T10:01:00Z"
...